Split IsStream.Nesting into Expand and Reduce
